Changed the manner in which PlayerLoopInterface cleans up added systems.
The old solution of just resetting the player loop back to the default was always a bit bad, since it assumed that every use of PlayerLoop went through the interface. It got extra bad when Unity started using the PlayerLoop internally, and assuming that the inserted system stayed in there forever. See:
https://forum.unity.com/threads/update-to-1-1-1-multi-binding-actions-no-longer-working.1172420/#post-7575136
https://github.com/Unity-Technologies/InputSystem/pull/1424/files#diff-9e93fa9ce6927588c3ffd63c65e117f9aa8879e106f379f0a1d1f19db79a40d0R125

The new solution is probably the one that should have been there from the start - the PlayerLoopInterface keeps track of all the types it has inserted, and tries to remove all of them.